move 'TrackChangesItem' to separate file
